Add endpoint to retrieve a list of agents in JSON format that have the given phone number
Add AgentsController#list_with_phone that returns an array of Agents
in JSON format that have the given phone number. An empty JSON array
([]) is returned if no results were found.

An SQLite3 user function, match_phone_number, is used
(/config/initializers/database_functions.rb) to normalize
the agent.phone_number against the phone number provided using
regex, in order to perform matching. TODO: Perhaps a better long-term
option might be to create a migration or rake task to strip formatting,
and normalize the existing phone numbers in the database; this, in
conjunction with stripping any formatting and normalizing the
Agent.phone_number prior to committing the model (before_validation?).