Add Policy requisite AgentPolicy model and controller/action to create policies via post
Policies cannot be created if they are not issuable; see
PolicyIssuable concern that provides a method that can
be used as a check to see if the policy can be issued
by the agent. In hindsight, the "is issuable" validation
logic currently in the Policy model should be moved to
the PolicyIssuable concern where it belongs.

In addition to the above, sorry for the big commit, it's been
a rough couple of weeks :S

Add JSON serialization support using ActiveModel::Serializers
(see /app/serializers) to make JSON rending in the controller
actions easier. Supporting modules can be found in /app/modules.
Serializers needed to handle collections of Agent models
are coded at the controller action level; this processing can
be moved to the Agent model as a class method and be called
to clean things up a little in the controller if necessary
later on.

Phone number matching against inconsistent phone number formats
in the database (agent.phone_number) to "get
/api/agents?phone_number=<numeric>" was painful. A database
function was added to normalize the phone numbers on the fly
to handle this, see
config/initializers/database_functions.rb. Not very efficient,
but the next best thing to updating the database and Agent
model to format the phone number on create and update.

Add appropriate test files for test coverate for the aforementioned.